Create Extension to convert a Response to a Failure -> expose to other object
