+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/network/Request.kt
@@ -16,24 +16,15 @@
 
 package im.vector.matrix.android.internal.network
 
-import com.squareup.moshi.JsonDataException
-import com.squareup.moshi.Moshi
-import im.vector.matrix.android.api.failure.ConsentNotGivenError
 import im.vector.matrix.android.api.failure.Failure
-import im.vector.matrix.android.api.failure.MatrixError
-import im.vector.matrix.android.internal.di.MoshiProvider
 import kotlinx.coroutines.CancellationException
-import org.greenrobot.eventbus.EventBus
 import retrofit2.Call
-import retrofit2.Response
-import timber.log.Timber
 import java.io.IOException
 
 internal suspend inline fun <DATA> executeRequest(block: Request<DATA>.() -> Unit) = Request<DATA>().apply(block).execute()
 
 internal class Request<DATA> {
 
-    private val moshi: Moshi = MoshiProvider.providesMoshi()
     lateinit var apiCall: Call<DATA>
 
     suspend fun execute(): DATA {
@@ -54,31 +45,5 @@ internal class Request<DATA> {
                 else                        -> Failure.Unknown(exception)
             }
         }
-    }
-
-    private fun <T> Response<T>.toFailure(): Failure {
-        val errorBody = errorBody() ?: return Failure.Unknown(RuntimeException("errorBody() should not be null"))
-
-        val errorBodyStr = errorBody.string()
-
-        val matrixErrorAdapter = moshi.adapter(MatrixError::class.java)
-
-        try {
-            val matrixError = matrixErrorAdapter.fromJson(errorBodyStr)
-
-            if (matrixError != null) {
-                if (matrixError.code == MatrixError.M_CONSENT_NOT_GIVEN && !matrixError.consentUri.isNullOrBlank()) {
-                    // Also send this error to the bus, for a global management
-                    EventBus.getDefault().post(ConsentNotGivenError(matrixError.consentUri))
-                }
-
-                return Failure.ServerError(matrixError, code())
-            }
-        } catch (ex: JsonDataException) {
-            // This is not a MatrixError
-            Timber.w("The error returned by the server is not a MatrixError")
-        }
-
-        return Failure.OtherServerError(errorBodyStr, code())
     }
 }
