+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/network/RetrofitExtensions.kt
@@ -18,14 +18,22 @@
 
 package im.vector.matrix.android.internal.network
 
+import com.squareup.moshi.JsonDataException
+import im.vector.matrix.android.api.failure.ConsentNotGivenError
+import im.vector.matrix.android.api.failure.Failure
+import im.vector.matrix.android.api.failure.MatrixError
+import im.vector.matrix.android.internal.di.MoshiProvider
 import kotlinx.coroutines.suspendCancellableCoroutine
+import okhttp3.ResponseBody
+import org.greenrobot.eventbus.EventBus
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
+import timber.log.Timber
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 
-suspend fun <T> Call<T>.awaitResponse(): Response<T> {
+internal suspend fun <T> Call<T>.awaitResponse(): Response<T> {
     return suspendCancellableCoroutine { continuation ->
         continuation.invokeOnCancellation {
             cancel()
@@ -40,4 +48,46 @@ suspend fun <T> Call<T>.awaitResponse(): Response<T> {
             }
         })
     }
+}
+
+/**
+ * Convert a retrofit Response to a Failure, and eventually parse errorBody to convert it to a MatrixError
+ */
+internal fun <T> Response<T>.toFailure(): Failure {
+    return toFailure(errorBody(), code())
+}
+
+/**
+ * Convert a okhttp3 Response to a Failure, and eventually parse errorBody to convert it to a MatrixError
+ */
+internal fun okhttp3.Response.toFailure(): Failure {
+    return toFailure(body(), code())
+}
+
+private fun toFailure(errorBody: ResponseBody?, httpCode: Int): Failure {
+    if (errorBody == null) {
+        return Failure.Unknown(RuntimeException("errorBody should not be null"))
+    }
+
+    val errorBodyStr = errorBody.string()
+
+    val matrixErrorAdapter = MoshiProvider.providesMoshi().adapter(MatrixError::class.java)
+
+    try {
+        val matrixError = matrixErrorAdapter.fromJson(errorBodyStr)
+
+        if (matrixError != null) {
+            if (matrixError.code == MatrixError.M_CONSENT_NOT_GIVEN && !matrixError.consentUri.isNullOrBlank()) {
+                // Also send this error to the bus, for a global management
+                EventBus.getDefault().post(ConsentNotGivenError(matrixError.consentUri))
+            }
+
+            return Failure.ServerError(matrixError, httpCode)
+        }
+    } catch (ex: JsonDataException) {
+        // This is not a MatrixError
+        Timber.w("The error returned by the server is not a MatrixError")
+    }
+
+    return Failure.OtherServerError(errorBodyStr, httpCode)
 }
